Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,12 +1,12 @@
 #include <stdio.h>
 int main() 
 {
-    float a,b,c;
+    int a,b,c;
     float res;
     scanf("%d",&a);
     scanf("%d",&b);
     scanf("%d",&c);
-    res=(a+b+c)/3;
+    res=(float)(a+b+c)/3;
     printf("Average: %.2f",res);
     return 0;
 }
